refactor: move remove_spin_projection
to temporary nested_dict module
